requirements.txt
langchain
langchain-groq
sqlmodel
fastapi
uvicorn
pydantic-settings
python-dotenv
langchain_community